requirements.txt
Django>=4.2
python-barcode>=0.15.1
Pillow>=10.0.0
openpyxl>=3.1.2
pandas>=2.0.0
reportlab>=4.0.0
qrcode>=7.4.2
